Auto include through associations when 'middle' association was loaded but not modified (#138)
When an association goes through another association, e.g. `has_many :tags, through: :post_tags`,
loading the `post_tags` association was enough, to not auto include the `tags` association when
it is first accessed.
This is correct for when the `post_tags` association was modified and auto include could lead to
weird results. When nothing was modified it is totally fine to auto include the through association.